Update number of internal variety production change
Before in one round only one internal trade was made. This was making
the system slower to stabilize and in the wrong scale. Now the first
always happen, the second happen with probability 70%, the third have
again a 70% an so on. In this aditional internal trades if the variety
with the lower delta no longer exists a random one will be picket to be
replaced.