Update variety definition limits
The definition of variety considers a limit of 0.04 in quality and 0.2
in half saturation constant. This second value is similar to the
variance of the initical distribution of half saturation constant.
Now this limit is smaller for half saturation (0.1) and a little bigger
for quality (0.05). The way the variety number was calculated have been
updated and also the max number of varieties changed.