Adds Parameter::key and is_valid_key
Instead of verifying in the main.cpp if the argument is any of the valid
char, which is hard to read and maintain. It created a Parameter::key
static vector containing all the valid keys. And a is_valid_key function
that return true a given character is in the Parameter::key vector.