Fix saving wrong timeline data on Results
There was a mismatch in the timeline data. Only the number of varieties
in the community was been written in the beggining and only the
punctuations were been written in the end. This commit fix it, now all
timeline data is written in the beggining and during the simulation. To
make this possible now the computeHDpunctuations is runned on household
initialization, this way the varietyOwened is filled and the household
metrics can be computed before running a iteration.