Creates methods to add results inside Result class
The methods that push data to the Result members were inside the Model
class. This is not right scope for it and there were repeated code
inside each tipe of model. Now the Result has a Metrics instance in it
and all the methods that pushes data to Results members using the
Metrics methods are inside the Result class. This way Model class does
not have to handle Metrics methods anymore.