Fix missing case in Parameter::set_parameter
One of the program modes ('m') is for making multiple model runs
disagragated. In this case the ModelRunner::run_var_param is used, but
with nRun=1 and the same parameters in each executions. To make this
possible this commit adds an 'm' option in the Parameter::set_parameter,
in this case no parameter is updated.