Fix model begin and end time conditions
Beginning the main time loop of the model with t = 0 was making the
first iteration to always call Result::save_timeline. This was causing a
shift in the resultant csv. Now the main time loop starts in t = 1 and
the second timeline result is only written when t = timeInterval. Also
this commit makes t = maxTime be in the result data. Before the last
time being written was t = maxTime - timeInterval.