fix path to write files
error: after the program finished running no data was being saved.

cause: previously the code were being compiled and runned inside the src
folder, this way the path where to write data was only 'test', now with
the change to compile the code with makefiles, it is running from the
project root, so now the path must be different.

solution: make base path be 'src/test'.